tons of changes made to sounds and minigolf
lots of sound adjustments, fixed ball rspawn so it actually updates location, added visible ball reset marker for each ball, cahnged texture of of courses to make them look more interesting but also to help with seeing issues in blender designs, added gameobjects for holes for balls to collide with when they go in and respawn at the initial position, added sounde effects to baseball for hitting the ball and foul balls and homeruns, adjusted skeleton position, added randomization to pitch speeds so now they can go mildly faster or slower as well as a bit higher or lower, added failsafe incase ball hits player diamond and is thus too low to reasonably hit the count will not increment and an additional pitch will be thrown, adjusted positioning of several buttons in menu as well as button s in basketball for ease of reading and use for everyone, added music that plays for duration of basketball game and a little after it ends. Other stuff that I forgot about too.